requirements.txt
pyserial>=3.5
readchar>=4.2.1
fastapi>=0.110
uvicorn[standard]>=0.29
